# [DO NOT MERGE] core: switch Bn256Mul precompile to evmone