Prerequisite: https://github.com/erigontech/evmone_precompiles/pull/1

BEFORE
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkPrecompiledBn256ScalarMul$ github.com/erigontech/erigon/core/vm


goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/core/vm
cpu: Apple M2 Max
BenchmarkPrecompiledBn256ScalarMul/chfast1-Gas=6000-12             25912         46088 ns/op          6000 gas/op          130.2 mgas/s     1280 B/op         25 allocs/op
BenchmarkPrecompiledBn256ScalarMul/chfast2-Gas=6000-12             23901         50040 ns/op          6000 gas/op          119.9 mgas/s     1472 B/op         26 allocs/op
BenchmarkPrecompiledBn256ScalarMul/chfast3-Gas=6000-12             24816         48238 ns/op          6000 gas/op          124.4 mgas/s     1472 B/op         26 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio1-Gas=6000-12            23089         51227 ns/op          6000 gas/op          117.1 mgas/s     1504 B/op         26 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio6-Gas=6000-12            23217         51276 ns/op          6000 gas/op          117.0 mgas/s     1504 B/op         26 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio11-Gas=6000-12           23320         51604 ns/op          6000 gas/op          116.3 mgas/s     1504 B/op         26 allocs/op
PASS
ok      github.com/erigontech/erigon/core/vm    10.872s
```

AFTER
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkPrecompiledBn256ScalarMul$ github.com/erigontech/erigon/core/vm

goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/core/vm
cpu: Apple M2 Max
BenchmarkPrecompiledBn256ScalarMul/chfast1-Gas=6000-12         	   56040	     21167 ns/op	      6000 gas/op	       283.4 mgas/s	      64 B/op	       1 allocs/op
BenchmarkPrecompiledBn256ScalarMul/chfast2-Gas=6000-12         	   24253	     50114 ns/op	      6000 gas/op	       119.7 mgas/s	      64 B/op	       1 allocs/op
BenchmarkPrecompiledBn256ScalarMul/chfast3-Gas=6000-12         	   22110	     55195 ns/op	      6000 gas/op	       108.7 mgas/s	      64 B/op	       1 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio1-Gas=6000-12        	   21633	     55445 ns/op	      6000 gas/op	       108.2 mgas/s	      64 B/op	       1 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio6-Gas=6000-12        	   21849	     54347 ns/op	      6000 gas/op	       110.4 mgas/s	      64 B/op	       1 allocs/op
BenchmarkPrecompiledBn256ScalarMul/cdetrio11-Gas=6000-12       	   22059	     54936 ns/op	      6000 gas/op	       109.2 mgas/s	      64 B/op	       1 allocs/op
PASS
ok  	github.com/erigontech/erigon/core/vm	10.863s
```