@yperbasis 2025-06-19
Closing in favour of #15556